+++ crux-rocksdb/src/crux/kv/rocksdb/metrics.clj
@@ -17,30 +17,24 @@
                                                        :ticker-count (.getAndResetTickerCount stats ticker-type)}))))]
                         bin-size))
 
-(def metrics-module
-  {::metrics {:start-fn (fn [{:keys [crux.node/kv-store crux.metrics/registry]} {::keys [bin-size shutdown-timeout]}]
-                          (let [meters (->> (seq (TickerType/values))
-                                            (into {} (map (fn [^TickerType ticker-type]
-                                                            [ticker-type (dw/meter registry ["rocksdb" (ticker-type-name ticker-type)])]))))
-                                stats (get-in kv-store [:kv :stats])
-                                collector (doto (->collector stats
-                                                             (fn [{:keys [ticker-type ticker-count]}]
-                                                               (some-> (get meters ticker-type) (dw/mark! ticker-count)))
-                                                             bin-size)
-                                            .start)
-                                num-snapshots (dw/gauge registry ["rocksdb" "num-snapshots"] #(.getLongProperty (get-in kv-store [:kv :db]) "rocksdb.num-snapshots"))]
-                            {:rocks-meters meters
-                             :num-snapshots num-snapshots}
+(defn start-rocksdb-metrics
+  [{:keys [crux.node/kv-store crux.metrics/registry crux.kv.rocksdb/metrics?]} {::keys [bin-size shutdown-timeout]}]
 
-                            (reify Closeable
-                              (close [_]
-                                (.shutDown collector shutdown-timeout)))))
+  (when metrics?
+    (let [meters (->> (seq (TickerType/values))
+                      (into {} (map (fn [^TickerType ticker-type]
+                                      [ticker-type (dw/meter registry ["rocksdb" (ticker-type-name ticker-type)])]))))
+          stats (get-in kv-store [:kv :stats])
+          collector (doto (->collector stats
+                                       (fn [{:keys [ticker-type ticker-count]}]
+                                         (some-> (get meters ticker-type) (dw/mark! ticker-count)))
+                                       bin-size)
+                      .start)
+          num-snapshots (dw/gauge registry ["rocksdb" "num-snapshots"]
+                                  #(.getLongProperty (get-in kv-store [:kv :db]) "rocksdb.num-snapshots"))]
+      {:rocks-meters meters
+       :num-snapshots num-snapshots}
 
-              :deps #{:crux.node/kv-store :crux.metrics/registry}
-              :before #{:crux.metrics/all-metrics-loaded}
-              :args {::bin-size {:doc "Sample size of statistics collector in milliseconds"
-                                 :default 3000
-                                 :crux.config/type :crux.config/int}
-                     ::shutdown-timeout {:doc "Time for statistics collector object to finish collecting in milliseconds"
-                                         :default 5000
-                                         :crux.config/type :crux.config/int}}}})
+      (reify Closeable
+        (close [_]
+          (.shutDown collector shutdown-timeout))))))